Crate zng_wgt_button

Button widget.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 1 feature flag, enabled by default.

"tooltip"

Enable tooltip in cmd buttons.

Enabled by default.

Structs

BUTTON

Button context.

Button

W A clickable container.

CmdTooltipArgs

Arguments for cmd_tooltip_fn.

DefaultStyle

W Button default style.

LightStyle

W Button light style.

LinkStyle

W Button link style.

PrimaryStyle

W Primary button style.

Statics

CMD_CHILD_FN_VAR

Widget function used when cmd is set and child is not.

CMD_PARAM_VAR

Optional parameter for the button to use when notifying command.

CMD_TOOLTIP_FN_VAR

Widget function used when cmd is set and tooltip_fn, tooltip are not set.

STYLE_FN_VAR

Contextual style variable.

Functions

cmd

c Sets the Command the button represents.

cmd_child_fn

P Sets the widget function used to produce the button child when cmd is set and child is not.

cmd_param

P Optional command parameter for the button to use when notifying cmd.

cmd_tooltip_fn

P Sets the widget function used to produce the button tooltip when cmd is set and tooltip is not.

default_cmd_child_fn

Default CMD_CHILD_FN_VAR.

default_cmd_tooltip_fn

Default CMD_TOOLTIP_FN_VAR.

style_fn

P Extends or replaces the widget style.