Crate zng_wgt_button

Button widget.

Crate

This crate is part of the zng project.

Structs

BUTTON

Button context.

Button

W A clickable container.

CmdTooltipArgs

Arguments for cmd_tooltip_fn.

DefaultStyle

W Button default style.

LinkStyle

W Button link style.

Statics

BASE_COLORS_VAR

Idle background dark and light color.

CMD_CHILD_FN_VAR

Widget function used when cmd is set and child is not.

CMD_PARAM_VAR

Optional parameter for the button to use when notifying command.

CMD_TOOLTIP_FN_VAR

Widget function used when cmd is set and tooltip_fn, tooltip are not set.

STYLE_FN_VAR

Contextual style variable.

Functions

base_colors

P Sets the colors used to compute all background and border colors in the button style.

cmd

c Sets the Command the button represents.

cmd_child_fn

P Sets the widget function used to produce the button child when cmd is set and child is not.

cmd_param

P Optional command parameter for the button to use when notifying cmd.

cmd_tooltip_fn

P Sets the widget function used to produce the button tooltip when cmd is set and tooltip is not.

color_scheme_hovered

Create a color_scheme_highlight of 0.08.

color_scheme_pressed

Create a color_scheme_highlight of 0.16.

default_cmd_child_fn

Default CMD_CHILD_FN_VAR.

default_cmd_tooltip_fn

Default CMD_TOOLTIP_FN_VAR.

style_fn

P Extends or replaces the widget style.